fix: create job in DB only when needed (#48)
when creating a job and directly setting the owner of it, `setOwner`
adds the job to the specified User, too. As the User is persisted on
every action, this triggered a save of the Job on any request to the job
API.

This fixes this in a way which still allows to assume a `job` instance
variable in the controller. But before saving the job, the owner and
creation date needs to be set explicitely.
What it entails is a few more checks for validating that we actually
found a job on the job API.