fix: don't count on orchestrator returning a valid return value (#76)
Choria orchestrator uses `shell_exec()` which returns NULL if there's an
error or if the command yielded no output. Otherwise a string.
The ssh commands executed always return null as output because of
`>/dev/null 2>&1` in the command. So checking that return value,
especially in the form of removing manager nodes as returning something
truthy is wrong.