fix: callback remove job handling (#93)
Remove job worked but the following callback call by Choria had an issue
because Choria just sends the node name and not as assumed before the
node FQDN when deleting. This is different from the creation of a node,
where Choria sends the node's FQDN.

Additionally this fixes several issues with the callback handling and
the conditions. This was introduced with the new feature of having a
separate manager entity and the code is a bit convoluted as a result
because of checking for the deprecated status etc.
We might want to get rid of that stuff soon in order to clean up our
codebase.

The code about reprovisioning the manager within the callback is most
likely not needed anymore and would not do anything anyway at this
point. Therefore removed this bit. We might need to revisit how all that
works together later.