fix: sliding window setting of next active execution (#134)
There are two cases where we increase number of replicas on another
execution:
- running execution is done -> calls submitresult -> schedule next
execution
- new worker in cluster -> increase number of running executions

Previous behaviour worked for both cases only on the "current" job,
meaning the one the API got called with.
When submitresult was called it checked for other jobs to run as well
(as a secondary thought basicly).
When new worker in cluster it would work only if the current job still
has executions not running yet.

New behaviour is it just fetches all runnable executions which belong to
a job matching the same labels. It orders them by creation date so
earlier ones go first. We also sort by priority of the job - in case we
ever need/want to prioritise jobs.

In theory this could mean that job A and B run at the same time as soon
as a new worker gets added to the cluster (in case A and B have been
created +/- at the same time).